implementación currencyService

+++ src/Service/CategoryService.php
@@ -4,12 +4,14 @@ namespace App\Service;
 
 use App\Repository\CategoryRepository;
 use App\Entity\Category;
+use App\Repository\ProductRepository;
 
 class CategoryService
 {
-    public function __construct(CategoryRepository $categoryRepository)
+    public function __construct(CategoryRepository $categoryRepository, ProductRepository $productRepository)
     {
         $this->categoryRepository = $categoryRepository;
+        $this->productRepository = $productRepository;
     }
 
     public function getCategoryById(int $id) :?Category
@@ -19,8 +21,7 @@ class CategoryService
 
     public function saveCategory(Category $category) :bool
     {
-        $this->categoryRepository->saveCategory($category);
-        return true;
+        return $this->categoryRepository->saveCategory($category);
     }
 
     public function updateCategory(int $id, Category $category) :bool
@@ -29,9 +30,7 @@ class CategoryService
         if ($currentCategory != null) {
             $currentCategory->setName($category->getName());
             $currentCategory->setDescription($category->getDescription());
-            $this->saveCategory($currentCategory);
-            
-            return true;
+            return $this->saveCategory($currentCategory);
         }
 
         return false;
@@ -41,8 +40,12 @@ class CategoryService
     {
         $currentCategory = $this->getCategoryById($id);
         if ($currentCategory != null) {
+            $productsOfCategory = $this->productRepository->getProductsByCategoryId($id);
+            if ($productsOfCategory != null) {
+                return false;
+            }
+
             return $this->categoryRepository->removeCategory($currentCategory);
-            return true;
         }
 
         return false;
